Rename slices directory
